pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "astro-stack"
version = "0.1.0"
description = "Automated astrophotography processing pipeline with event-driven API"
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
authors = [{ name = "Astro-Stack Team" }]

dependencies = [
    # Web framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "websockets>=12.0",
    # Database
    "sqlmodel>=0.0.19",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",
    "greenlet>=3.0.0",
    # Task queue
    "arq>=0.26.0",
    # Redis pub/sub
    "redis[hiredis]>=5.0.0",
    # Config
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
    # Auth
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    # HTTP async (Ollama)
    "httpx>=0.27.0",
    # Astrophotography
    "astropy>=6.0.0",
    "rawpy>=0.22.0",
    "numpy>=1.26.0",
    "Pillow>=10.3.0",
    # File watching
    "watchdog>=4.0.0",
    # Utilities
    "python-multipart>=0.0.9",
    "aiofiles>=23.2.0",
    "anyio>=4.4.0",
    "structlog>=24.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "httpx>=0.27.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
    "types-passlib",
    "types-python-jose",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

# ---------------------------------------------------------------------------
# Ruff — lint + format (replaces black + isort + flake8)
# ---------------------------------------------------------------------------
[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "ANN", "B", "SIM", "RUF"]
ignore = [
    "ANN101",  # Missing type annotation for `self`
    "ANN102",  # Missing type annotation for `cls`
    "ANN401",  # Dynamically typed expressions (Any) allowed
    "B008",    # Do not perform function calls in default arguments (FastAPI Depends)
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["ANN", "S101"]
"alembic/*" = ["ANN", "INP001"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false

# ---------------------------------------------------------------------------
# Mypy
# ---------------------------------------------------------------------------
[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true

# ---------------------------------------------------------------------------
# Pytest
# ---------------------------------------------------------------------------
[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "--cov=app --cov-report=term-missing --cov-fail-under=0"

[tool.coverage.run]
source = ["app"]
omit = ["tests/*", "alembic/*"]
